smartfridge/backend/requirements.txt
Django>=5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
drf-spectacular>=0.27
python-dotenv>=1.0
psycopg[binary]>=3.2
celery>=5.3 ; python_version>='3.9'
redis>=5.0 ; python_version>='3.9'
dj-database-url>=2.1
openai>=1.51.2
